Improve Concatenate Click handling #56
Sometimes when the user clicks on an unchecked checkbox a click event is
sent, but the box is not flagged as checked. Had to improve the code to
filter out these clicks because when the DataGrid items are recreated it
was adding the item the unclicked list twice and causing an exception.