Adding Raster field name to list of excluded fields #83
If source and target datasets both have a raster field Append will work.
Not sure if we want to add this to future requirements, DA will not
perform any processing on the fields, it's more a question of whether
the Append/Replace functions should populate these values in the target
datasets.