Ensure DefaultValues do not override config settings, make xml files portable
Fixes to make xml files portable between machines #112, the code now
stores a relative path to the project or the Xml config file. This
should support running processes from other projects as long as the
config file is stored in the same (parent) folder as the dataset. Also
made changes to address #130 to ensure that DefaultValues from source
datasets do not override config settings - this could happen when there
are source and target fields with the same name and a defaultValue on
the source field.