Resolve issue with Append failing #128
There was a copy and paste error in some code, it was getting the wrong
field type in some cases. Also made a small fix to the DefaultValue code
#130 for fields like GlobalID that cannot have a defaultValue. Also
reduced the number of print statement when there is a large number of
errors, this problem will occur since we removed the limit for max error
count that used to be in the tools.